Break plugin/manager apart
